requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.61.0
pydantic==2.10.5
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.4
